client/agent/requirements.txt
# OpsRamp AI Agent Requirements
aiohttp>=3.8.0
aiohttp-sse-client>=0.2.1
requests>=2.28.0

# LLM Providers (at least one is required)
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.3.0

# Utilities
asyncio
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
